feat: streamline ui, add contextless snackbar presentation, model fix

+++ lib/domain/models/employee_edit/employee_edit.dart
@@ -15,7 +15,7 @@ class EmployeeEdit with _$EmployeeEdit {
   @override
   final String name;
   @override
-  final double salary;
+  final int salary;
   @override
   final int age;
